started work on refactoring code and introducing interface for specific recorders
